BIOVEC/requirements.txt
numpy
scikit-learn
biopython
gensim
torch
skorch
